Refactor tests into separate files and remove classes
- I split tests from `tests/test_tools.py` into:
  - `tests/test_add_text.py` (for `add_text` functionality)
  - `tests/test_check_weblinks.py` (for `check_weblinks` functionality)
- `tests/test_tools.py` now only contains tests for `get_page_content`.
- I removed class-based test grouping in favor of top-level pytest functions.
- I ensured consistent use of `pytest-mock`'s `mocker` fixture, including for `MagicMock`.

Note: I attempted Ruff formatting steps but was repeatedly interrupted. The structural changes are the primary focus of this commit.
